cleaned up search and trying to solve PV

+++ src/engine/inc/search/search_constants.hpp
@@ -21,5 +21,5 @@ static constexpr i32 c_maxScore = 32000;
 static constexpr i32 c_checkmateConstant = 24000;
 static constexpr i32 c_checkmateMaxDistance = 256;
 static constexpr i16 c_checkmateMinScore = c_checkmateConstant - c_checkmateMaxDistance;
-static constexpr i32 c_drawConstant = 0;
+static constexpr i16 c_drawConstant = 0;
 //static constexpr i32 c_pvScore = 10000;
